+++ mercury/mercury-library/models/src/main/scala/io/iohk/atala/mercury/DidComm.scala
@@ -9,6 +9,7 @@ trait DidComm {
   def myDid: DidId // TODO
   def packSigned(msg: Message): UIO[SignedMesage]
   def packEncrypted(msg: Message, to: DidId): UIO[EncryptedMessage]
+  def packEncryptedForward(msg: Message, to: DidId): UIO[EncryptedMessage]
   def unpack(str: String): UIO[UnpackMessage]
   def unpackBase64(dataBase64: String): UIO[UnpackMessage] = {
     val data = new String(Base64.getUrlDecoder.decode(dataBase64))
@@ -23,6 +24,9 @@ object DidComm {
     ZIO.serviceWithZIO(_.packSigned(msg))
 
   def packEncrypted(msg: Message, to: DidId): URIO[DidComm, EncryptedMessage] =
+    ZIO.serviceWithZIO(_.packEncrypted(msg, to))
+
+  def packEncryptedForward(msg: Message, to: DidId): URIO[DidComm, EncryptedMessage] =
     ZIO.serviceWithZIO(_.packEncrypted(msg, to))
 
   def unpack(str: String): URIO[DidComm, UnpackMessage] =